chore: Move the WordArea type to its own module
